+++ src/openApi/v3/Parser.ts
@@ -1,5 +1,4 @@
 import { Context } from '../../core/Context';
-import { getRefFromSchema } from '../../utils/getRefFromSchema';
 import { getModel } from '../v3/parser/getModel';
 import { getModelComposition } from './parser/getModelComposition';
 import { getModelProperties } from './parser/getModelProperties';